+++ src/GmGM/typing.py
@@ -1,5 +1,8 @@
+from __future__ import annotations
+
 import numpy as np
 import scipy.sparse as sparse
+from typing import TypeVar, Union
 
 try:
     # TypeAlias only added in Python 3.10
@@ -11,10 +14,18 @@ Axis: TypeAlias = str
 Modality: TypeAlias = str
 DataTensor: TypeAlias = np.ndarray
 
+
+_key = TypeVar('_key')
+_value = TypeVar('_value')
+
+# | only added in Python 3.10 :(
+MaybeDict = Union[_value, dict[_key, _value]]
+
 """
 In python 3.12, this becomes:
 
 type Axis = str
 type Modality = str
 type DataTensor = np.ndarray
+type MaybeDict[Key, Value] = Value | dict[Key, Value]
 """